requirements.txt
PyQt5==5.14.2
python-dotenv==0.13.0
requests==2.23.0
